return early when counting words for multi-draft if no path found
I originally skipped the scene if the scenePath wasn't found, but scene paths are only `null` if the folder itself can't be found.  If that's the case, we'll get a bunch of duplicate error logs, one for each scene.  Instead, we avoid even iterating the scenes if the scene folder can't be found.